Augment UDP send failure warning.
If sending an event over UDP fails, the failure notice that is sent now
has a timestamp and default fields, just like the original. This makes
it easier to see where a failure event occured (e.g. when one of the
default fields is the hostname).

This also changes `augment` to modify the passed event dictionary
instead of starting a new one and copying over all values.